Refactor add-action-no-template
